extract components from stockCard and change stockType to union type
