+++ src/assets/sass/lbd/_tables.scss
@@ -59,10 +59,14 @@
     }
 }
 
-.small-padding-left {
+.stockDetails {
   padding-left: 3em;
 }
 
 .small-padding-top {
   padding-top: 0.5em;
+}
+
+.small-padding-left {
+  padding-left: 1em;
 }
