+++ src/views/Market/StockCard/index.ts
@@ -1,0 +1,3 @@
+import StockCard from './StockCard';
+
+export default StockCard;
